Install VisualVM libraries during the build
The OpenIDE/VisualVM libraries are not available from Maven central. Add
the VisualVM ZIP file to the source tree and create POMs to install the
required libraries during the build so that we don't need to rely on
external Maven repositories. This is a prerequisite to be able to deploy
releases to Maven central.